Sixth commit "AccountService, TransactionService, GoalService, AccountController, TransactionController, GoalController"

+++ src/main/java/com/pocketplanner/model/Account.java
@@ -23,7 +23,7 @@ public class Account {
     private String name;
 
     @Column(name = "balance", nullable = false)
-    private BigDecimal balance;
+    private Double balance;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
